+++ src/constants/workflows.js
@@ -3,4 +3,12 @@ const WORKFLOW_TYPES = [
   {value: 'Opportunity', text: 'Opportunity'}
 ];
 
+const NAME_PREFIXES = [
+  {value: 'Mr', text: 'Mr'},
+  {value: 'Mrs', text: 'Mrs'},
+  {value: 'Dr', text: 'Dr'},
+  {value: 'Sir', text: 'Sir'},
+];
+
+export {NAME_PREFIXES};
 export default WORKFLOW_TYPES