+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+## [0.30.2] - 2020-04-27
+### Changed
+- Page 'Create New Requester'
+    - select "prefix" -> input"prefix". 
+
 ## [0.30.1] - 2020-04-27
 ### Added 
 - Page 'Create New Requester'